First working triangle purely from Rust bindings.
Added more partial bindings, and got the first fully rust-driven
triangle rendering to the screen.

+++ cpp/src/view.cc
@@ -14,9 +14,9 @@ extern "C" void View_SetCamera(View* view, Camera* camera) {
   view->setCamera(camera);
 }
 
-extern "C" void View_SetViewport(View* view, int left, int bottom, int width,
-                                 int height) {
-  view->setViewport({left, bottom, (uint32_t)width, (uint32_t)height});
+extern "C" void View_SetViewport(View* view, int32_t left, int32_t bottom,
+                                 uint32_t width, uint32_t height) {
+  view->setViewport({left, bottom, width, height});
 }
 
 extern "C" void View_SetClearColor(View* view, float linearR, float linearG,
@@ -37,7 +37,8 @@ extern "C" void View_SetClearTargets(View* view, bool color, bool depth,
   view->setClearTargets(color, depth, stencil);
 }
 
-extern "C" void View_SetVisibleLayers(View* view, int select, int value) {
+extern "C" void View_SetVisibleLayers(View* view, int32_t select,
+                                      int32_t value) {
   view->setVisibleLayers((uint8_t)select, (uint8_t)value);
 }
 
@@ -49,42 +50,42 @@ extern "C" void View_SetRenderTarget(View* view, RenderTarget* renderTarget) {
   view->setRenderTarget(renderTarget);
 }
 
-extern "C" void View_SetSampleCount(View* view, int count) {
+extern "C" void View_SetSampleCount(View* view, int32_t count) {
   view->setSampleCount((uint8_t)count);
 }
 
-extern "C" int View_GetSampleCount(View* view) {
+extern "C" int32_t View_GetSampleCount(View* view) {
   return view->getSampleCount();
 }
 
-extern "C" void View_SetAntiAliasing(View* view, int type) {
+extern "C" void View_SetAntiAliasing(View* view, int32_t type) {
   view->setAntiAliasing(View::AntiAliasing(type));
 }
 
-extern "C" int View_GetAntiAliasing(View* view) {
+extern "C" int32_t View_GetAntiAliasing(View* view) {
   return (int)view->getAntiAliasing();
 }
 
-extern "C" void View_SetToneMapping(View* view, int type) {
+extern "C" void View_SetToneMapping(View* view, int32_t type) {
   view->setToneMapping(View::ToneMapping(type));
 }
 
-extern "C" int View_GetToneMapping(View* view) {
+extern "C" int32_t View_GetToneMapping(View* view) {
   return (int)view->getToneMapping();
 }
 
-extern "C" void View_SetDithering(View* view, int dithering) {
+extern "C" void View_SetDithering(View* view, int32_t dithering) {
   view->setDithering((View::Dithering)dithering);
 }
 
-extern "C" int View_GetDithering(View* view) {
+extern "C" int32_t View_GetDithering(View* view) {
   return (int)view->getDithering();
 }
 
 extern "C" void View_SetDynamicResolutionOptions(
     View* view, bool enabled, bool homogeneousScaling,
     float targetFrameTimeMilli, float headRoomRatio, float scaleRate,
-    float minScale, float maxScale, int history) {
+    float minScale, float maxScale, int32_t history) {
   View::DynamicResolutionOptions options;
   options.enabled = enabled;
   options.homogeneousScaling = homogeneousScaling;
@@ -97,7 +98,8 @@ extern "C" void View_SetDynamicResolutionOptions(
   view->setDynamicResolutionOptions(options);
 }
 
-extern "C" void View_SetRenderQuality(View* view, int hdrColorBufferQuality) {
+extern "C" void View_SetRenderQuality(View* view,
+                                      int32_t hdrColorBufferQuality) {
   View::RenderQuality renderQuality;
   renderQuality.hdrColorBuffer = View::QualityLevel(hdrColorBufferQuality);
   view->setRenderQuality(renderQuality);
@@ -108,7 +110,7 @@ extern "C" void View_SetDynamicLightingOptions(View* view, float zLightNear,
   view->setDynamicLightingOptions(zLightNear, zLightFar);
 }
 
-extern "C" void View_SetDepthPrepass(View* view, int value) {
+extern "C" void View_SetDepthPrepass(View* view, int32_t value) {
   view->setDepthPrepass(View::DepthPrepass(value));
 }
 
@@ -128,11 +130,11 @@ extern "C" bool View_IsFrontFaceWindingInverted(View* view) {
   return static_cast<bool>(view->isFrontFaceWindingInverted());
 }
 
-extern "C" void View_SetAmbientOcclusion(View* view, int ordinal) {
+extern "C" void View_SetAmbientOcclusion(View* view, int32_t ordinal) {
   view->setAmbientOcclusion((View::AmbientOcclusion)ordinal);
 }
 
-extern "C" int View_GetAmbientOcclusion(View* view) {
+extern "C" int32_t View_GetAmbientOcclusion(View* view) {
   return (int)view->getAmbientOcclusion();
 }
